changed void init() to bool init()

+++ AndroidGatewayPlugin/main.cpp
@@ -60,7 +60,7 @@ private:
   ~App();
 
 public:
-  void init(int argc, char* argv[]);
+  bool init(int argc, char* argv[]);
   void run();
   void stop();
 
@@ -115,7 +115,7 @@ App::~App()
   }
 }
 
-void App::init(int argc, char* argv[])
+bool App::init(int argc, char* argv[])
 {
   dropPrivileges();
   setupLogging("AndroidGatewayPlugin");
@@ -156,7 +156,7 @@ void App::init(int argc, char* argv[])
       LOG_FATAL("                           interface (default 32869)");
       LOG_FATAL("  --listenAddress address  Sets the listening address for the Android");
       LOG_FATAL("                           interface (default 0.0.0.0, or all interfaces)");
-      return;
+      return false;
     }
   }
   
@@ -171,6 +171,8 @@ void App::init(int argc, char* argv[])
 								             androidPort);
 
   LOG_INFO("Listening on port " << androidPort << " on interface " << androidAddress);
+
+  return true;
 }
 
 void App::run()
@@ -189,9 +191,9 @@ void App::stop()
 
 #ifdef WIN32
 
-void SvcInit(DWORD argc, LPTSTR* argv)
+bool SvcInit(DWORD argc, LPTSTR* argv)
 {
-  App::instance()->init(argc, argv);
+  return App::instance()->init(argc, argv);
 }
 
 void SvcRun()
@@ -245,7 +247,9 @@ int main(int argc, char* argv[])
 #else
 int main(int argc, char** argv)
 {
-  App::instance()->init(argc, argv);
+  if (!App::instance()->init(argc, argv)) {
+    return 1;
+  }
   App::instance()->run();
   App::instance()->destroy();
   return 0;
