+++ GatewayCore/main.cpp
@@ -62,7 +62,7 @@ private:
   ~App();
 
 public:
-  void init(int argc, char* argv[]);
+  bool init(int argc, char* argv[]);
   void run();
   void stop();
 
@@ -117,7 +117,7 @@ App::~App()
   }
 }
 
-void App::init(int argc, char* argv[])
+bool App::init(int argc, char* argv[])
 {
   dropPrivileges();
   setupLogging("GatewayCore");
@@ -148,6 +148,8 @@ void App::init(int argc, char* argv[])
   
   //Initializes the cross-gateway connections
   GatewayCore::getInstance()->initCrossGateway();
+
+  return true;
 }
 
 void App::run()
@@ -167,9 +169,9 @@ void App::stop()
 
 #ifdef WIN32
 
-void SvcInit(DWORD argc, LPTSTR* argv)
+bool SvcInit(DWORD argc, LPTSTR* argv)
 {
-  App::instance()->init(argc, argv);
+  return App::instance()->init(argc, argv);
 }
 
 void SvcRun()
@@ -223,7 +225,9 @@ int main(int argc, char* argv[])
 #else
 int main(int argc, char** argv)
 {
-  App::instance()->init(argc, argv);
+  if (!App::instance()->init(argc, argv)) {
+    return 1;
+  }
   App::instance()->run();
   App::instance()->destroy();
   return 0;
